Nginx: Moved config file to /mnt
This is to resolve an error you receive when trying to start the container: "Error while creating mount source path '/var/nginx': mkdir /var/nginx: read-only file system." This issue did not happen when Docker was manually installed via apt but Google is suggesting this may happen when you install Docker as part of the Ubuntu OS Installation wizard.

Not really best practice to put it in Mount but only a homelab.